+++ src/corelib/text/qregexp.cpp
@@ -891,7 +891,7 @@ static bool operator==(const QRegExpEngineKey &key1, const QRegExpEngineKey &key
            && key1.cs == key2.cs;
 }
 
-static uint qHash(const QRegExpEngineKey &key, uint seed = 0) noexcept
+static size_t qHash(const QRegExpEngineKey &key, size_t seed = 0) noexcept
 {
     QtPrivate::QHashCombine hash;
     seed = hash(seed, key.pattern);
@@ -4041,7 +4041,7 @@ bool QRegExp::operator==(const QRegExp &rx) const
     Returns the hash value for \a key, using
     \a seed to seed the calculation.
 */
-uint qHash(const QRegExp &key, uint seed) noexcept
+size_t qHash(const QRegExp &key, size_t seed) noexcept
 {
     QtPrivate::QHashCombine hash;
     seed = hash(seed, key.priv->engineKey);
